Moving log functions into logger.py
